+++ modules/include/dxrt_drv.h
@@ -51,8 +51,10 @@
 /* RT/driver sync     */
 
 typedef enum {
+    DXRT_EVENT_NONE,
     DXRT_EVENT_ERROR,
     DXRT_EVENT_NOTIFY_THROT,
+    DXRT_EVENT_RECOVERY,
     DXRT_EVENT_NUM,
 } dxrt_event_t;
 
@@ -229,7 +231,7 @@ typedef enum {
     DXRT_CMD_READ_OUTPUT_DMA_CH0,
     DXRT_CMD_READ_OUTPUT_DMA_CH1,
     DXRT_CMD_READ_OUTPUT_DMA_CH2,
-    DXRT_CMD_TERMINATE          ,
+    DXRT_CMD_TERMINATE_EVENT    ,
     DXRT_CMD_EVENT              ,
     DXRT_CMD_DRV_INFO           , /* Sub-command */
     DXRT_CMD_SCHEDULE           , /* Sub-command */
@@ -240,6 +242,7 @@ typedef enum {
     DXRT_CMD_RECOVERY           ,
     DXRT_CMD_CUSTOM             , /* Sub-command */
     DXRT_CMD_START              ,
+    DXRT_CMD_TERMINATE          ,
     DXRT_CMD_MAX,
 } dxrt_cmd_t;
 
@@ -449,6 +452,7 @@ extern dxrt_message_handler message_handler[];
 #define dx_pcie_interrupt(...) 0
 #define dx_pcie_interrupt_clear(...) 0
 #define dx_pcie_interrupt_wakeup(...) 0
+#define dx_pcie_interrupt_event_wakeup(...) 0
 #define dx_pcie_dequeue_response(...) 0
 #define dx_pcie_get_dev_num(...) 1
 #define dx_pcie_get_download_region(...) 0