+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # RELEASE_NOTES
 
+## v1.5.0 / 2025-05-14
+### 1. Changed
+- None
+### 2. Fixed
+- None
+### 3. Added
+- Implements termiation api
+  If the user wants to terminate normally, the corresponding API is called and the driver is notified of termination.
+
 ## v1.4.3 / 2025-05-12
 ### 1. Changed
 - None